requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.15
botocore==1.27.15
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.3
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
